RatNet-PyTorch/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
nuscenes-devkit>=1.1.9
numpy>=1.19.2
Pillow>=8.0.0
PyYAML>=5.4.1
tensorboard>=2.4.0
tqdm>=4.50.0 